+++ appMain.py
@@ -621,7 +621,7 @@ class App(QtCore.QObject):
 
         self.app_units = self.options["units"]
         self.default_units = self.defaults["units"]
-        self.decimals = int(self.options['decimals_metric'])
+        self.decimals = int(self.options['units_precision'])
 
         if self.options["global_theme"] == 'default':
             self.resource_location = 'assets/resources'